+++ changelog.md
@@ -9,3 +9,11 @@
 - [command/tax] handle taxCode & taxRate AND add for each tax 2 types additive & inclusive @maramalshen
 - [action/invoice] handle TaxCodeRef in the invoice & return_invoice @maramalshen
 - [command/basic] fix bug: sync taxes before sync products @maramalshen
+
+### Changed
+
+### Fixed
+
+- [commands/customers] fix bug, get customers paginated from qp @maramalshen
+
+### Removed